requirements.dev.txt
pytest==6.2.0
pytest-aiohttp==0.3.0
Faker==5.0.2
numpy==1.19.4